pyproject.toml
# ====================================================================================
# Build System Configuration
# ====================================================================================
[build-system]
requires = [ "hatchling" ]
build-backend = "hatchling.build"

# ====================================================================================
# Project Metadata (PEP 621)
# ====================================================================================
[project]
name = "wallcat"
description = "Wallpaper Catalog & Classifier"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Heliomar P. Marques", email = "heliomarpm@proton.me" }]
classifiers = [
  "Programming Language :: Python :: 3.11",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
keywords = [
  "wallpaper",
  "image-classification",
  "cli",
  "automation",
  "computer-vision"
]
dynamic = ["version"]
dependencies = [
    "loguru>=0.7.3",
    "open-clip-torch>=3.2.0",
    "pillow>=12.1.0",
    "pyyaml>=6.0.3",
    "torch>=2.10.0",
    "typer>=0.21.1",
]

[project.optional-dependencies]
dev = [
  "ruff", "pyright", "pytest", "pytest-cov"
]

[project.scripts]
wallcat = "wallcat.main:main"  # executable: uv run wallcat classify ./wallpapers --mode hybrid

[project.urls]
Homepage = "https://navto.me/heliomarpm/"
"Repository" = "https://github.com/heliomarpm/wallcat"
"Documentation" = "https://github.com/heliomarpm/wallcat"
"Bug Tracker" = "https://github.com/heliomarpm/wallcat/issues"
"Changelog" = "https://github.com/heliomarpm/wallcat/releases"


# ====================================================================================
# Tool Configuration for Setuptools, Ruff, etc.
# For more information https://setuptools.pypa.io/en/latest/index.html
# ====================================================================================
[tool.hatch.version]
path = "src/wallcat/__init__.py"

# ---- Pytest Configuration ----------------------------------------------------------
[tool.pytest.ini_options]
addopts = "-s --color=yes --tb=short"
pythonpath = ["src"]
testpaths = ["tests"]


# ---- Pyright Configuration ----------------------------------------------------------
[tool.pyright]
typeCheckingMode = "standard"
# pythonVersion = "3.14"
include = ["src", "tests"]
exclude = ["**/venv", "**/.venv", "**/.env", "**/__pycache__"]
venv = ".venv"
venvPath = "."
executionEnvironments = [{ root = "src" }]

[tool.uv.sources]
clip = { git = "https://github.com/openai/CLIP.git" }
